fix: move emitter wrapper out one level, to catch user functions properly
